Add a scenario for testing resilience of Search to index churn
One of the known problems with Search that if indexes are switched
in a quick succession, the state shared between Dreyfus and
Clouseau gets out of sync, which will eventually lead to a stuck
instance and requires a restart.  This test is added to emulate
such cases and verify the node is sufficiently equipped to avoid
these problems.